agent_backend/requirements.txt
openai>=1.10.0
langgraph>=0.0.30
ollama>=0.2.0 # Added ollama dependency
Sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
python-dotenv>=1.0.0 # For config
numpy>=1.21.0      # Optional: For InMemoryVectorStore example

# Backend Core
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0     # For database migrations (requires Alembic)
Alembic>=1.9.0
Flask-RESTful>=0.3.9    # For structured REST APIs
Flask-JWT-Extended>=4.4.0 # For JWT authentication
passlib>=1.7.4          # For password hashing
itsdangerous>=2.0.0     # For token signing (dependency of Flask/JWT)
Marshmallow>=3.15.0     # For serialization/deserialization
Flask-SocketIO>=5.3.0   # For WebSocket support
eventlet>=0.33.0        # Or gevent - WebSocket server option for Flask-SocketIO
Flask-Caching>=2.0.0    # For caching support
# Add Redis if distributed caching is desired: redis>=4.3.0
requests>=2.20.0        # For test client login
python-socketio[asyncio_client]>=5.8.0 # For WebSocket client
bcrypt>=4.0.0